ACS-8262: Conditional test execution.
