+++ .secrets.baseline
@@ -157,6 +157,38 @@
         "is_secret": false
       }
     ],
+    "distribution/src/main/resources/docker-compose/docker-compose-minimal.yml": [
+      {
+        "type": "Secret Keyword",
+        "filename": "distribution/src/main/resources/docker-compose/docker-compose-minimal.yml",
+        "hashed_secret": "9e3d103f7aa5f4f778cf752087dfceeba15d4fef",
+        "is_verified": false,
+        "line_number": 12
+      },
+      {
+        "type": "Secret Keyword",
+        "filename": "distribution/src/main/resources/docker-compose/docker-compose-minimal.yml",
+        "hashed_secret": "cadf6fe64f4c39aacbb64df9a4894ba173ba2248",
+        "is_verified": false,
+        "line_number": 22
+      },
+      {
+        "type": "Secret Keyword",
+        "filename": "distribution/src/main/resources/docker-compose/docker-compose-minimal.yml",
+        "hashed_secret": "a4a747bd4ba5e3a5049cad116881867c71fb625b",
+        "is_verified": false,
+        "line_number": 62,
+        "is_secret": false
+      },
+      {
+        "type": "Secret Keyword",
+        "filename": "distribution/src/main/resources/docker-compose/docker-compose-minimal.yml",
+        "hashed_secret": "40f81ac66c7cb246ded7b83e15ca0bf5d7f57eee",
+        "is_verified": false,
+        "line_number": 146,
+        "is_secret": false
+      }
+    ],
     "distribution/src/main/resources/docker-compose/docker-compose-w-keycloak.yml": [
       {
         "type": "Secret Keyword",
@@ -294,5 +326,5 @@
       }
     ]
   },
-  "generated_at": "2024-06-21T14:12:40Z"
+  "generated_at": "2024-07-01T10:13:17Z"
 }
