+++ e2e-test/src/test/java/org/alfresco/hxi_connector/e2e_test/CreateNodeE2eTest.java
@@ -25,32 +25,15 @@
  */
 package org.alfresco.hxi_connector.e2e_test;
 
-import static com.github.tomakehurst.wiremock.client.WireMock.containing;
-import static com.github.tomakehurst.wiremock.client.WireMock.exactly;
-import static com.github.tomakehurst.wiremock.client.WireMock.moreThanOrExactly;
-import static com.github.tomakehurst.wiremock.client.WireMock.postRequestedFor;
-import static com.github.tomakehurst.wiremock.client.WireMock.urlEqualTo;
-import static org.assertj.core.api.Assertions.assertThat;
-
-import java.io.ByteArrayInputStream;
-import java.io.File;
 import java.io.IOException;
-import java.io.InputStream;
-import java.util.ArrayList;
-import java.util.List;
 
 import com.github.tomakehurst.wiremock.client.WireMock;
-import lombok.Cleanup;
 import lombok.SneakyThrows;
-import lombok.extern.slf4j.Slf4j;
-import org.apache.commons.collections4.CollectionUtils;
-import org.apache.pdfbox.Loader;
-import org.apache.pdfbox.io.RandomAccessReadBuffer;
-import org.apache.pdfbox.pdmodel.PDDocument;
-import org.apache.pdfbox.text.PDFTextStripper;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.TestInstance;
+import org.junit.jupiter.api.condition.DisabledIfEnvironmentVariable;
 import org.testcontainers.containers.BindMode;
 import org.testcontainers.containers.GenericContainer;
 import org.testcontainers.containers.Network;
@@ -62,22 +45,18 @@ import org.wiremock.integrations.testcontainers.WireMockContainer;
 
 import org.alfresco.hxi_connector.common.test.docker.repository.AlfrescoRepositoryContainer;
 import org.alfresco.hxi_connector.common.test.docker.util.DockerContainers;
-import org.alfresco.hxi_connector.common.test.util.RetryUtils;
 import org.alfresco.hxi_connector.e2e_test.util.client.AwsS3Client;
 import org.alfresco.hxi_connector.e2e_test.util.client.RepositoryNodesClient;
-import org.alfresco.hxi_connector.e2e_test.util.client.model.Node;
-import org.alfresco.hxi_connector.e2e_test.util.client.model.S3Object;
 
-@Slf4j
 @Testcontainers
 @SuppressWarnings({"PMD.FieldNamingConventions"})
-public class CreateNodeE2eTest
+@TestInstance(TestInstance.Lifecycle.PER_CLASS)
+@DisabledIfEnvironmentVariable(named = "GITHUB_ACTIONS_DEPLOY_TIMEOUT", matches = ".*")
+/**
+ * As of now this test class is excluded from GitHub Actions workflow but will run with the maven builds (thus, relies on GITHUB_ACTIONS_DEPLOY_TIMEOUT env variable being set).
+ */
+public class CreateNodeE2eTest extends CreateNodeE2eTestBase
 {
-    private static final String BUCKET_NAME = "test-hxinsight-bucket";
-    private static final int MAX_ATTEMPTS = 5;
-    private static final int INITIAL_DELAY_MS = 500;
-    private static final String PARENT_ID = "-my-";
-    private static final String DUMMY_CONTENT = "Dummy's file dummy content";
 
     private static final Network network = Network.newNetwork();
     @Container
@@ -104,13 +83,12 @@ public class CreateNodeE2eTest
     private static final AlfrescoRepositoryContainer repository = createRepositoryContainer()
             .dependsOn(postgres, activemq, transformCore, transformRouter, sfs, liveIngester);
 
-    RepositoryNodesClient repositoryNodesClient = new RepositoryNodesClient(repository.getBaseUrl(), "admin", "admin");
-    AwsS3Client awsS3Client = new AwsS3Client(awsMock.getHost(), awsMock.getFirstMappedPort(), BUCKET_NAME);
-
     @BeforeAll
     @SneakyThrows
-    public static void beforeAll()
+    public void beforeAll()
     {
+        repositoryNodesClient = new RepositoryNodesClient(repository.getBaseUrl(), "admin", "admin");
+        awsS3Client = new AwsS3Client(awsMock.getHost(), awsMock.getFirstMappedPort(), BUCKET_NAME);
         WireMock.configureFor(hxInsightMock.getHost(), hxInsightMock.getPort());
         awsMock.execInContainer("awslocal", "s3api", "create-bucket", "--bucket", BUCKET_NAME);
     }
@@ -125,60 +103,14 @@ public class CreateNodeE2eTest
     @SuppressWarnings({"PMD.JUnitTestsShouldIncludeAssert"})
     void testCreateNodeContainingImageFile() throws IOException
     {
-        // given
-        File imageFile = new File("src/test/resources/images/quick.jpg");
-        List<S3Object> initialBucketContent = awsS3Client.listS3Content();
-
-        // when
-        Node createdNode = repositoryNodesClient.createNodeWithContent(PARENT_ID, imageFile);
-
-        // then
-        RetryUtils.retryWithBackoff(() -> {
-            List<S3Object> actualBucketContent = awsS3Client.listS3Content();
-            assertThat(actualBucketContent.size()).isEqualTo(initialBucketContent.size() + 1);
-
-            WireMock.verify(exactly(1), postRequestedFor(urlEqualTo("/presigned-urls")));
-            WireMock.verify(moreThanOrExactly(2), postRequestedFor(urlEqualTo("/ingestion-events"))
-                    .withRequestBody(containing(createdNode.id())));
-        }, MAX_ATTEMPTS, INITIAL_DELAY_MS);
+        testCreateNodeContainingImageFileBase();
     }
 
     @Test
     @SuppressWarnings({"PMD.JUnitTestsShouldIncludeAssert"})
     void testCreateNodeContainingTextFile() throws IOException
     {
-        // given
-        @Cleanup
-        InputStream fileContent = new ByteArrayInputStream(DUMMY_CONTENT.getBytes());
-        List<S3Object> initialBucketContent = awsS3Client.listS3Content();
-
-        // when
-        Node createdNode = repositoryNodesClient.createNodeWithContent(PARENT_ID, "dummy.txt", fileContent, "text/plain");
-
-        // then
-        RetryUtils.retryWithBackoff(() -> {
-            List<S3Object> actualBucketContent = awsS3Client.listS3Content();
-            assertThat(actualBucketContent.size()).isEqualTo(initialBucketContent.size() + 1);
-
-            S3Object s3Object = new ArrayList<>(CollectionUtils.disjunction(initialBucketContent, actualBucketContent)).get(0);
-            String actualPdfContent = getPdfContent(s3Object.key());
-            assertThat(actualPdfContent).isEqualToIgnoringWhitespace(DUMMY_CONTENT);
-
-            WireMock.verify(exactly(1), postRequestedFor(urlEqualTo("/presigned-urls")));
-            WireMock.verify(moreThanOrExactly(2), postRequestedFor(urlEqualTo("/ingestion-events"))
-                    .withRequestBody(containing(createdNode.id())));
-        }, MAX_ATTEMPTS, INITIAL_DELAY_MS);
-    }
-
-    @SneakyThrows
-    private String getPdfContent(String objectKey)
-    {
-        @Cleanup
-        InputStream pdfContent = awsS3Client.getS3ObjectContent(objectKey);
-        @Cleanup
-        PDDocument document = Loader.loadPDF(new RandomAccessReadBuffer(pdfContent));
-        PDFTextStripper pdfStripper = new PDFTextStripper();
-        return pdfStripper.getText(document);
+        testCreateNodeContainingTextFileBase();
     }
 
     private static AlfrescoRepositoryContainer createRepositoryContainer()
